Refactor: Split send-email-smtp function
Split the send-email-smtp function into smaller files for better maintainability.
